% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate spatial weights for points&/polygons}
\usage{
calc_weights(tweets, distance, d1, d2)
}
\arguments{
\item{tweets}{SpatialPointsDataFrame containing tweets}

\item{distance}{distance metrics}

\item{d1}{fixed distance minimum}

\item{d2}{fixed distance maximum}
}
\description{
Calculate spatial weights for points&/polygons
}
