% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getisord.R
\name{getisOrd}
\alias{getisOrd}
\title{Perform Getis-Ord}
\usage{
getisOrd(tweets, distance, d1, d2)
}
\arguments{
\item{tweets}{data to be analyzed}

\item{distance}{distance metrics}

\item{d1}{fixed distance minimum}

\item{d2}{fixed distance maximum}
}
\description{
Perform Getis-Ord
}
