% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello2}
\alias{hello2}
\title{Hello, world2!}
\usage{
hello2()
}
\description{
This is an example function named 'hello'
which prints 'Hello, world2!'.
}
\details{
Some useful keyboard shortcuts for package authoring:

Build and Reload Package:  'Ctrl + Shift + B'
  Check Package:             'Ctrl + Shift + E'
  Check Package:             'Ctrl + Shift + E'
  Test Package:              'Ctrl + Shift + T'
}
