% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise}
\alias{visualise}
\title{Visualise - create map USA}
\usage{
visualise(referencedregion, tweets)
}
\arguments{
\item{referencedregion}{area of interest}

\item{tweets}{data to work with}
}
\description{
Visualise - create map USA
}
