% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdata.R
\name{read.data}
\alias{read.data}
\title{Function read.data is reading the dataset to be used in the analysis}
\usage{
read.data(fileName)
}
\arguments{
\item{fileName}{name of the file}
}
\description{
Function read.data is reading the dataset to be used in the analysis
}
