% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{read.data}
\alias{read.data}
\title{Function read.data is reading the dataset to be used in the analysis}
\usage{
read.data(fileName)
}
\description{
Function read.data is reading the dataset to be used in the analysis
}
